requirements.txt
Flask
Flask-CORS
selenium
pillow
pytesseract
